refactor(v010): carve voicegw serve out of _legacy
Ninth per-command carve-out in section 2.

Adds voicegateway/cli/serve.py owning the 'serve' command. Body
unchanged from v0.0.5; the lazy 'import uvicorn' stays inside the
function body so the cli does not eagerly load the optional
dashboard extras when the user is just running, say, voicegw status.

13 commands still register on the shared app; pytest tests/test_cli.py
46/46 pass; ruff and mypy clean across the cli package (10 source
files now).
